+++ src/types/league.ts
@@ -1,0 +1,8 @@
+import { Team } from "./team";
+
+export interface SoccerLeague {
+  name: string;
+  teams: Team[];
+  enabled: boolean;
+  flagSVGcode: string;
+}
