+++ src/types/team.ts
@@ -1,0 +1,4 @@
+export interface Team {
+  name: string;
+  code: string;
+}
